Update the core generic canonicalization facility
The core canonicalization facility has been significantly updated.  Now
the experimental path class is significantly augmented and handles the
addition of candidates for both the cases with automorphism group and
the cases without automorphism group.  The usage of a hash map for both
cases significantly improved the code clarity reduced duplication.