Reverse multiplication order in Schreier Lemma
At the core of the Schreier-Sims algorithm, the Schreier Lemma is
utilized to get the generators of the subgroup from generators of the
full group and coset representatives.  The form of the lemma on
Wikipedia and most books are for right cosets, but our application
focuses on left cosets.  In this commit, the order of multiplication are
all reversed for get the analogous lemma.

Note that this is just based on intuition, rather than some rigorous
proof.