Reorganize action of perm on string objects
Now a general templates function is added for generic action of perm on
string objects, with the result able to be tuned for any valid result
type.  This makes the code able to be reused in test cases and user
cases.

Also some additional cosmetic formatting and revision is made.