Fix the discovery of a new automorphism
The automorphisms to be put into the transversal system are the ones
that should give the children when it is left-multiplied to the anchor.
The previous code does the opposite.  Now it is fixed and commented.

Also an assertion is added for the new insert semantics of the
transversal concept.