Change when refinement is performed in Eldag canon
In the canonicalization of Eldags, previously the refinement is
performed on the testing of leaves.  This causes big problem with the
const-ness of the cosets in the canonicalization process.

Now this code has been revised by performing the refinement after the
initialization.  This could make the coset to be constant as required by
the canonicalization facility.