Make iteration on individualized eldag cosets lazy
Previously, the individualized cosets are first put into a vector
eagerly.  Then it is mandated that we copy construct its entries in the
actual hash map for them.  This has not only detrimental effects on the
performance, and it requires unnecessary code for the copy construction
as well.

Now this is fixed by making the result of refinement to be a lazy
iterable.  It should have very high performance.