Add facility to correct cell iterators after splitting
Currently partition cell iterators is heavily used while the partition
itself is being refined during the iteration process.  In this process,
when a cell is shattered, the iterators pointing to the cell will point
to a **non-deterministic** cell from the shattered cell.  In this
commit, facilities are added to correct cell iterators when cells are
split by a given key function, based on the direction of the iterators.