Fix bug in transversal system adaptation
In commit 62a6d4ad, some optimization was introduced to the transversal
system adaption algorithm, where the product of the permutation with
identity, i.e. the permutation itself, is attempted to be treated before
its product with other coset representatives in the system, in hope to
be able to move some of these permutations.  However, if some of them
are indeed moved, the permutations will be deprived of its actual
content and left in an undefined state.  Although the code tries to
update the pointers to new locations after the move, when the
permutation is actually in a coset with an existing representative, the
pointer will still refer to a permutation in this broken state.  When it
used in the later product expression, the program will crash with a
segmentation fault.

In spite of a chance that such permutations can be ignored for later
treatment, I cannot find a rigorous proof of this, even after some
serious trials.  As a result, for better safety, this optimization is
completely removed, with the code falling back to a more naive
treatment.  Since the possible saving is at most on the scale of O(n),
while the algorithm itself has a worse case complexity of O(n^4), the
elimination of this optimization is not expected to pose a discernible
regression in the performance of the code.