Fix bug in transversal system adaptation better
In commit 943529d, a bug in the transversal adaptation algorithm causing
segfault has been investigated and attempted to be fixed.  However,
although the fix of removing an optimization works, the rationale is not
the actual cause of the issue.  In the pre-loop, the permutations cannot
be moved away with the pointer not updated.  The optimization is
conceptually correct.  However, after we move some permutations into the
vector of permutations to pass for the next loop, we simply store a
pointer to the back of the vector.  When the vector later changes its
capacity because of newly added permutations, the pointer will be
invalidated and points to undefined bytes.

In this commit, the issue is tried to be fixed with this understanding.
The optimization of moving the permutations first is kept.  But the swap
vector is moved away from inside the loop, and is reserved the maximum
possible space it can possibly use.  The inside this loop, we swap this
swap space with the space for the actual permutations to pass.  In this
way, since the permutations to pass will monotonically decrease during
the loop, it is guaranteed that the capacity will not change, also we
are bound to have only two memory allocations, making the code possibly
more efficient than before.